requirements.txt
Flask==3.0.3
Werkzeug==3.0.2
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.30
WTForms==3.1.2
alembic==1.13.1
boto3==1.34.108
botocore==1.34.108
python-dotenv==1.0.1